pyproject.toml
[tool.poetry]
name = "skel"
version = "0.1.0"
description = "Skeleton for a Python project"
authors = ["Jonathan González <jgonf@safebytelabs.com>"]
license = "MPL2"
readme = "README.md"

# -- manually added - begin of block ----------------------
#
packages = [{ include = "mypkg", from = "src" }]

[tool.semantic_release]
branch = "main"
version_variables = ["src/mypkg/__init__.py:__version__"]

[tool.pytest.ini_options]
pythonpath = ["src"]
#
# -- manually added - end of block ------------------------

[tool.black]
include = '\.pyi?$'
line-length = 120
skip-string-normalization = true

#
# poetry export -f requirements.txt --output requirements.txt --without-hashes
# awk '{ print $1 }' requirements.txt > temp.txt && mv temp.txt requirements.txt
#
[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
marshmallow = "^3.21.3"
requests = "^2.32.3"
redis = "^5.0.7"
flask-limiter = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
